+++ crates/masm-analysis/src/unconstrained_advice/shared.rs
@@ -1,20 +1,20 @@
 //! Shared transfer helpers for unconstrained-advice analyses.
 
+use masm_decompiler::{LoopPhi, Stmt, Var};
 pub(super) use masm_decompiler::{
-    INTRINSIC_ADV_PIPE, INTRINSIC_ADV_PUSH, INTRINSIC_ADV_PUSHW, INTRINSIC_MEM_STREAM,
-    intrinsic_base_name, intrinsic_memory_address_arg_index, intrinsic_merkle_root_arg_range,
+    intrinsic_memory_address_arg_index, intrinsic_merkle_root_arg_range,
     intrinsic_nonzero_arg_index, intrinsic_positional_u32_arg_range,
     intrinsic_requires_u32_precondition,
 };
-use masm_decompiler::{Intrinsic, LoopPhi, Stmt, Var};
 
-use super::{domain::AdviceFact, u32_domain::U32Validity};
+use super::domain::AdviceFact;
 pub(super) use super::{
     env::Env,
     expr::{
         assign_expr_metadata, collect_expr_sink_fact, expr_is_proven_nonzero, expr_output_fact,
         expr_u32_validity, refine_if_envs,
     },
+    intrinsic_transfer::{apply_intrinsic_effect, refine_nonzero_from_intrinsic},
     local_transfer::{
         apply_local_load_scalar, apply_local_load_word, apply_local_store, apply_local_store_word,
     },
@@ -137,104 +137,6 @@ pub(super) fn stabilized_loop_head_env(
     .env()
 }
 
-/// Refine the environment after `assertz` proves an `eq.0` witness is zero.
-pub(super) fn refine_nonzero_from_intrinsic(intrinsic: &Intrinsic, env: &mut Env) {
-    if intrinsic_base_name(&intrinsic.name) != "assertz" {
-        return;
-    }
-    let Some(arg) = intrinsic.args.first() else {
-        return;
-    };
-    let Some(witness) = env.zero_test_for_var(arg) else {
-        return;
-    };
-    env.mark_identity_nonzero(witness.value_identity);
-}
-
-/// Apply the common provenance transfer semantics of one intrinsic statement.
-pub(super) fn apply_intrinsic_effect(
-    span: miden_debug_types::SourceSpan,
-    intrinsic: &Intrinsic,
-    env: &mut Env,
-) {
-    if matches!(intrinsic_base_name(&intrinsic.name), INTRINSIC_ADV_PUSH | INTRINSIC_ADV_PUSHW) {
-        for result in &intrinsic.results {
-            env.set_var_fact(result, AdviceFact::from_source(span));
-            env.clear_var_metadata(result);
-        }
-        return;
-    }
-
-    match intrinsic_base_name(&intrinsic.name) {
-        "u32assert" | "u32assert2" | "u32assertw" => {
-            for arg in &intrinsic.args {
-                env.sanitize_var(arg);
-            }
-        },
-        "u32split" => {
-            for (index, result) in intrinsic.results.iter().enumerate() {
-                env.set_var_fact(result, AdviceFact::bottom());
-                env.clear_var_metadata(result);
-                if index == intrinsic.results.len().saturating_sub(1) {
-                    env.set_var_u32_validity(result, U32Validity::ProvenU32);
-                }
-            }
-        },
-        "is_odd" => {
-            for result in &intrinsic.results {
-                env.set_var_fact(result, AdviceFact::bottom());
-                env.clear_var_metadata(result);
-            }
-        },
-        "u32testw" => {
-            if let Some((flag, preserved)) = intrinsic.results.split_first() {
-                env.set_var_fact(flag, AdviceFact::bottom());
-                env.clear_var_metadata(flag);
-                env.set_var_u32_validity(flag, U32Validity::ProvenU32);
-                for (result, arg) in preserved.iter().zip(intrinsic.args.iter()) {
-                    env.set_var_fact(result, env.fact_for_var(arg));
-                    env.set_var_u32_validity(result, env.u32_validity_for_var(arg));
-                    env.set_var_identity(result, env.identity_for_var(arg));
-                    env.set_var_zero_test(result, env.zero_test_for_var(arg));
-                }
-            }
-        },
-        INTRINSIC_ADV_PIPE => {
-            apply_adv_pipe_effect(span, intrinsic, env);
-        },
-        INTRINSIC_MEM_STREAM => {
-            apply_mem_stream_effect(intrinsic, env);
-        },
-        "sdepth" => {
-            for result in &intrinsic.results {
-                env.set_var_fact(result, AdviceFact::bottom());
-                env.clear_var_metadata(result);
-            }
-        },
-        name if name.starts_with("locaddr") => {
-            for result in &intrinsic.results {
-                env.set_var_fact(result, AdviceFact::bottom());
-                env.clear_var_metadata(result);
-            }
-        },
-        _ if intrinsic_requires_u32_precondition(&intrinsic.name) => {
-            for result in &intrinsic.results {
-                env.set_var_fact(result, AdviceFact::bottom());
-                env.clear_var_metadata(result);
-                env.set_var_u32_validity(result, U32Validity::ProvenU32);
-            }
-        },
-        _ => {
-            let joined =
-                AdviceFact::join_all(intrinsic.args.iter().map(|arg| env.fact_for_var(arg)));
-            for result in &intrinsic.results {
-                env.set_var_fact(result, joined.clone());
-                env.clear_var_metadata(result);
-            }
-        },
-    }
-}
-
 /// Return the statement span for a structured statement.
 pub(super) fn stmt_span(stmt: &Stmt) -> miden_debug_types::SourceSpan {
     match stmt {
@@ -255,65 +157,5 @@ pub(super) fn stmt_span(stmt: &Stmt) -> miden_debug_types::SourceSpan {
         | Stmt::If { span, .. }
         | Stmt::While { span, .. }
         | Stmt::Return { span, .. } => *span,
-    }
-}
-
-/// Apply the custom transfer semantics of `mem_stream`.
-fn apply_mem_stream_effect(intrinsic: &Intrinsic, env: &mut Env) {
-    if intrinsic.args.len() == 13 && intrinsic.results.len() == 13 {
-        let address = &intrinsic.args[12];
-        env.set_var_fact(&intrinsic.results[0], env.fact_for_var(address));
-        env.set_var_u32_validity(&intrinsic.results[0], U32Validity::Unknown);
-        env.clear_var_metadata(&intrinsic.results[0]);
-
-        for (offset, result) in intrinsic.results[1..5].iter().enumerate() {
-            let preserved_input = &intrinsic.args[11 - offset];
-            env.set_var_fact(result, env.fact_for_var(preserved_input));
-            env.set_var_u32_validity(result, env.u32_validity_for_var(preserved_input));
-            env.set_var_identity(result, env.identity_for_var(preserved_input));
-            env.set_var_zero_test(result, env.zero_test_for_var(preserved_input));
-        }
-
-        for result in &intrinsic.results[5..] {
-            env.set_var_fact(result, AdviceFact::bottom());
-            env.clear_var_metadata(result);
-        }
-        return;
-    }
-
-    for result in &intrinsic.results {
-        env.set_var_fact(result, AdviceFact::bottom());
-        env.clear_var_metadata(result);
-    }
-}
-
-/// Apply the custom transfer semantics of `adv_pipe`.
-fn apply_adv_pipe_effect(
-    span: miden_debug_types::SourceSpan,
-    intrinsic: &Intrinsic,
-    env: &mut Env,
-) {
-    if intrinsic.args.len() == 13 && intrinsic.results.len() == 13 {
-        env.set_var_fact(&intrinsic.results[0], AdviceFact::bottom());
-        env.clear_var_metadata(&intrinsic.results[0]);
-
-        for (offset, result) in intrinsic.results[1..5].iter().enumerate() {
-            let preserved_input = &intrinsic.args[11 - offset];
-            env.set_var_fact(result, env.fact_for_var(preserved_input));
-            env.set_var_u32_validity(result, env.u32_validity_for_var(preserved_input));
-            env.set_var_identity(result, env.identity_for_var(preserved_input));
-            env.set_var_zero_test(result, env.zero_test_for_var(preserved_input));
-        }
-
-        for result in &intrinsic.results[5..] {
-            env.set_var_fact(result, AdviceFact::from_source(span));
-            env.clear_var_metadata(result);
-        }
-        return;
-    }
-
-    for result in &intrinsic.results {
-        env.set_var_fact(result, AdviceFact::from_source(span));
-        env.clear_var_metadata(result);
     }
 }
